Allow passing env file as a parameter.
With the addition of network mode and fwd mode as command line params,
there is no additional host config required. So reusing the file to
override env vars. This is useful for cases where you want to point to a
private repo - change images names etc as these are currently hardcoded
for a release.