Fix 57: Allow specifying ProxyCommand ansible option
This requires escaping the ans_opts variable as the syntax for
ProxyCommand requires a quote. ProxyCommand spec requires ncat to be
installed in the container. So creating a new version of the ansible
container and separating it from the installer build.
Now we can build and release the ansible container separately.
Updating the sed statements for env.json update to handle the case where
there are pre-existing values. This is required to repeat a failed
install.