Use system proxy
The intention of override this env is to provide envrionment vars
in many places of ansible roles (which might be a bad idea), but
force it to use no proxy will ignore whatever has been set on system
level and if there's requirement to do http/https proxy on the host,
it will not be able to reach the internet or other network depends
on the proxy settings. We use use system proxy by default, or override
the proxy settings explicitly if it's required.

Signed-off-by: Wei Tie <wtie@cisco.com>