Build legacy swarm with any netplugin branch
During make build, user can specify:
* NETPLUGIN_BRANCH - the branch to compile instead of latest release
* NETPLUGIN_OWNER (if not contiv) - the username for the netplugin fork

When the branch is specified, only the full installer is built.

A lot of the environment / config for the build has moved to the
Makefile and enviornment variables so it's easy to find and set, and
they are largely prefixed with CONTIV_

Some decomposition of the build has happened (ansible and netplugin
split out into separate scripts and targets), getting us part way to
build.sh being only an assembling of assets gathered by other scripts
and dropped in artifact_staging directory.

Preparing the release tarball is done in a random temp directory that
is cleaned up automatically.

build.sh no longer takes a bunch of CLI arguments, instead relies on
enviornment variables, as this script is intended to be run by the
Makefile.

The legacy swarm mode prefers the full installer vs. the over the
internet installer.

Created a reusable installer unpacker that later can be used by other
install scripts.

Drive-by:
- update the netplugin versions to 1.1.5