Format env.json as JSON format template
In JSON spec, a "true" with quotes is not a boolean but a string,
there was a lot sed doing line replacement because we passed in the
env.json as string instead of a file. However, ansible supports
"-e @file" which allows to pass-in a JSON file, and this commit
is to use that way. In details, it
1. format env.json to regular JSON format
2. put flag for sed replacement, and change line replacement in
install.sh and uninstall.sh to do string replacement.
3. change to use 4 spaces as indentation in bash script
4. add -x for install and uninstall script

Signed-off-by: Wei Tie <wtie@cisco.com>